Implemented ISubsystem interface.
Core now services subsystems, instead of creating frames for the worlds.
Worlds themselves are now subsystems. This implementation is a lot
cleaner and flexible, and means Core doesn't have to worry about what
it's actually servicing. (Now possible to only run a series of
subsystems, with no Worlds, for example)